requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
openai==1.6.1
orjson==3.9.10
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
Pygments==2.17.2
pymongo==4.8.0
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
